TSK-719: Upgrade from Junit4 to Junit5
This commit adds Junit5 dependencies to taskana-core and partially
updates tests in the pro.taskana.impl package to use Junit5.
Tests which use custom runners aren't updateable because the RunWith
annotation is ommited in Junit5. Furthermore, all tests that use the
Powermock library can't be upgraded because Junit5 is not supported by
the current Powermock version.